+++ examples/data-sources/nxos_feature_telnet/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_feature_telnet" "example" {
+}