+++ examples/resources/nxos_feature_telnet/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_telnet.example "sys/fm/telnet"