+++ examples/resources/nxos_feature_telnet/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_telnet" "example" {
+  admin_state = "enabled"
+}